+++ test/dummy/db/migrate/20150924203053_add_timespan.rb
@@ -1,0 +1,7 @@
+class AddTimespan < ActiveRecord::Migration
+
+  def change
+    add_column :things, :duration_in_seconds, :integer
+  end
+
+end