+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 Features:
 
   - Add SubmitBar control
+  - Add Timespan control